Remove none core components/fragments/utils
Remove all site specific things from core package.  Should leave that in site specific spots to figure out what filters to pass and how to display facet layouts.  Can look at adding this to the default theme if need be for global use, but this should only worry about getting info on how to query algolia and what values to pass back to the site.