Allow input of defaultWithSearchQuery + sortField.
This allows a particular template using this component to effectively set the default selected sort (the one following a search term being entered).

defaultWithSearchQuery defaults to the current internal default of Published if not present and if a sortField is not pass into the component it defaults to the current internal default of Published as well.