Add source to login & authenticate components and routes
This will set souce as the event name in the login component(login||comment-stream-login||content-meter-login) on login submit.  This is passed to sendAppUserLoginLink mutation and set as a query param for souce on the authenticate link.  It then can be read on the authenticate route/template and set in the additional event date object.