Strip oly_enc_id from query param URL.
This only happens on production in not incognito windows, I can’t figure out a way of testing this without just putting it on production.

I seemed to recall something about this being done deliberately, however on production in an incognito window the query parameter is properly stripped so I entirely don’t know what’s going on here.