Update default monorail site-header with icon-wrapper
Add new icon-wrapper element the now holds the icons that appear on the right.  At the moment it will contain search, newsletterToggler & site-menu-toggle(when positioned right).  I also added support for passing a menuBtnPosition prop left || right.  If left it display like it use to.  If right it will pull the menu toggler into the icon-wrapper as the last item.

alls added this setup to example site