Adjust comment stream css to account for comment count
Move max-height and hidding of overflow done from the wrapper to the actual comment stream.  This allows use to account for the where there are no comments or a single comment displaying the whole stream.

By default now the $comment-start-height ~ max-height will be applyed stream instead of the wrapper(stream + toggle button).  It will also not set it when there are no comments or only a single comment.

Also moved the show comments transition when the toggle button is pressed.