Add incremented node list support
Add support to latest content list block to pass modifiers through to the flow which if you add incremented-nodes will add the css styles to put incremented numbers in front of each node, like a ordered list.

At the moment the style is 1 ). This could be adjusted if need be.