Add gatedContent check to newsletterState middleware
Update the newsletterState middleware to check the incoming request for the presece of a content id/\d{8}/ off of request.params.id.  This will required the inclusion of this middleware to be moved from withing start-server to the individual route that this is need to be applied to.  This will also prevent this newsletter state from being falsely triggered on routes like __idx & other random backend routes.

I also adjusted when the setting of the coockie happens to only be when the initiallyExpanded value is true instead only checking to see if the coockie exsits.  No point in setting until we actually want to display it.