Add product.type.id to getOmedaCustomerRecord
This will be used in conjuntion with the onLoginLinkSentFormatter to determine if the user has already been subscribed to the current website product.

The example-website example is using this to filter down the subscriptions based on current rapidIdentificationProductId set by the config.  If the user is already subscribed to the website product remove any promoCode being set for the `registration_meter`. This promoCode, or associated promoCode, should have already been set the first time they registered and subscribed to the website.