Add profile onSubmit message and redirect logic
On submit hide the form and display a success message.  The message will be one of the two following message depending on if there is a returnTo url set or not.

WIth returnTo:
Your profile has been saved.
If you are not automatically redirected click here to continue.

Without returnTo:
Your profile has been saved.
Click here to continue modiying your profile or click here to return to the home page.