Remove ref to submit message & add scroll logic on submit
Remove old submitMessage logic as it will never show when the wrapping element is now in a v-if=“!didSubmit” check and the new messaging is moved to the else portion of that.

I also added logic to scroll to top of the page on submit to prevent the user from being shown the footer when the form hides.