Add function to only return transcript not empty
This will insure that the transcript field is actually set to something with text in it vs empty html.  Before if you cleared out or the user has just save a couple of empty html element it would return and display what looked to be nothing, but was really emply p  & h tags.  This check looks to see if there is actually text to render within the given html elements.