Add ability to force profile update on content access
At the moment it would just pass through the current content access payload to the /profile form submit.  Meaning it would still force the validation and submission of only data configured to be handdled.  Meaning in the case of FCP where they are asking specific demos per form & none are actually active on the profile, it will  not save them to the profile.