Setup initial cookie value & default cookieValue to 1
Set the default cookieValue to 1 and set and initialValue of 1.  Yes these value are seperate as some could mess with the incoming prop for whatever reason, but the initial value should always be 1.  Meaning if the cookie doesnt exist and you are not setting it the number of view should be 1.